scripts: fix all shellcheck warns & make all proposed changes

+++ build_clang.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
-# shellcheck disable=SC2086
 # A custom Multi-Stage LLVM Toolchain builder.
 # PGO optimized clang for building Linux Kernels.
 set -e
@@ -36,12 +35,12 @@ fi
 
 echo "Starting LLVM Build"
 
-rm -rf $KERNEL_DIR
-rm -rf $TEMP_BINTUILS_BUILD && mkdir -p $TEMP_BINTUILS_BUILD
-rm -rf $TEMP_BINTUILS_INSTALL && mkdir -p $TEMP_BINTUILS_INSTALL
+rm -rf "$KERNEL_DIR"
+rm -rf "$TEMP_BINTUILS_BUILD" && mkdir -p "$TEMP_BINTUILS_BUILD"
+rm -rf "$TEMP_BINTUILS_INSTALL" && mkdir -p "$TEMP_BINTUILS_INSTALL"
 
 if [[ $CLEAN_BUILD -eq 3 ]]; then
-    rm -rf $LLVM_BUILD
+    rm -rf "$LLVM_BUILD"
 fi
 
 # Where all relevant build-related repositories are cloned.
@@ -79,34 +78,34 @@ binutils_pull() {
 
 get_linux_5_tarball() {
 
-    if [ -e linux-$1.tar.xz ]; then
+    if [ -e linux-"$1".tar.xz ]; then
         echo "Existing linux-$1 tarball found, skipping download"
-        tar xf linux-$1.tar.xz
+        tar xf linux-"$1".tar.xz
     else
         echo "Downloading linux-$1 tarball"
         wget "https://cdn.kernel.org/pub/linux/kernel/v5.x/linux-$1.tar.xz"
-        tar xf linux-$1.tar.xz
+        tar xf linux-"$1".tar.xz
     fi
 }
 
 build_temp_binutils() {
 
-    rm -rf $TEMP_BINTUILS_BUILD
-    mkdir -p $TEMP_BINTUILS_BUILD
+    rm -rf "$TEMP_BINTUILS_BUILD"
+    mkdir -p "$TEMP_BINTUILS_BUILD"
     if [ "$1" = "aarch64-linux-gnu" ]; then
         USE_SYSTEM_BINUTILS_64=0
     else
         USE_SYSTEM_BINUTILS_32=0
     fi
-    cd $TEMP_BINTUILS_BUILD
+    cd "$TEMP_BINTUILS_BUILD"
     "$BINUTILS_DIR"/configure \
         CC="gcc" \
         CXX="g++" \
         CFLAGS="-march=x86-64 -mtune=generic -flto=auto -flto-compression-level=10 -O3 -pipe -ffunction-sections -fdata-sections" \
         CXXFLAGS="-march=x86-64 -mtune=generic -flto=auto -flto-compression-level=10 -O3 -pipe -ffunction-sections -fdata-sections" \
         LDFLAGS="-Wl,-O3,--sort-common,--as-needed,-z,now" \
-        --target=$1 \
-        --prefix=$TEMP_BINTUILS_INSTALL \
+        --target="$1" \
+        --prefix="$TEMP_BINTUILS_INSTALL" \
         --disable-compressed-debug-sections \
         --disable-gdb \
         --disable-gdbserver \
@@ -129,10 +128,10 @@ build_temp_binutils() {
         --quiet \
         --with-pkgversion="Neutron Binutils"
 
-    make -s -j$(nproc --all) >/dev/null
-    make install -s -j$(nproc --all) >/dev/null
+    make -s -j"$(nproc --all)" >/dev/null
+    make install -s -j"$(nproc --all)" >/dev/null
     echo "temp binutils build done, removing build dir"
-    rm -rf $TEMP_BINTUILS_BUILD
+    rm -rf "$TEMP_BINTUILS_BUILD"
 }
 
 bolt_profile_gen() {
@@ -159,7 +158,7 @@ bolt_profile_gen() {
             HOSTLD="$STAGE3"/ld.lld \
             CROSS_COMPILE=aarch64-linux-gnu-
 
-        perf record --output ${BOLT_PROFILES}/perf.data --event cycles:u --branch-filter any,u -- make all -s -j$(nproc --all) \
+        perf record --output "${BOLT_PROFILES}"/perf.data --event cycles:u --branch-filter any,u -- make all -s -j"$(nproc --all)" \
             LLVM=1 \
             LLVM_IAS=1 \
             ARCH=arm64 \
@@ -183,15 +182,15 @@ bolt_profile_gen() {
         cd "$OUT"
 
         "$STAGE1"/perf2bolt "$STAGE3"/clang-16 \
-            -p ${BOLT_PROFILES}/perf.data \
-            -o ${BOLT_PROFILES}/clang-16.fdata || (
+            -p "${BOLT_PROFILES}"/perf.data \
+            -o "${BOLT_PROFILES}"/clang-16.fdata || (
             echo "Failed to convert perf data"
             exit 1
         )
 
         "$STAGE1"/llvm-bolt "$STAGE3"/clang-16 \
             -o "$STAGE3"/clang-16.bolt \
-            --data ${BOLT_PROFILES}/clang-16.fdata \
+            --data "${BOLT_PROFILES}"/clang-16.fdata \
             -relocs \
             -split-functions \
             -split-all-cold \
@@ -214,7 +213,7 @@ bolt_profile_gen() {
         "$STAGE1"/llvm-bolt \
             --instrument \
             --instrumentation-file-append-pid \
-            --instrumentation-file=${BOLT_PROFILES}/clang-16.fdata \
+            --instrumentation-file="${BOLT_PROFILES}"/clang-16.fdata \
             "$STAGE3"/clang-16 \
             -o "$STAGE3"/clang-16.inst
 
@@ -242,7 +241,7 @@ bolt_profile_gen() {
             HOSTLD="$STAGE3"/ld.lld \
             CROSS_COMPILE=aarch64-linux-gnu-
 
-        make all -s -j$(nproc --all) \
+        make all -s -j"$(nproc --all)" \
             LLVM=1 \
             LLVM_IAS=1 \
             ARCH=arm64 \
@@ -265,8 +264,8 @@ bolt_profile_gen() {
         )
         cd "$OUT"
 
-        cd $BOLT_PROFILES
-        "$STAGE1"/merge-fdata *.fdata >combined.fdata
+        cd "$BOLT_PROFILES"
+        "$STAGE1"/merge-fdata ./*.fdata >combined.fdata
 
         "$STAGE1"/llvm-bolt "$STAGE3"/clang-16.org \
             --data combined.fdata \
@@ -290,16 +289,16 @@ bolt_profile_gen() {
 }
 
 if [ -d "$LLVM_DIR"/ ]; then
-    cd $LLVM_DIR/
+    cd "$LLVM_DIR"/
     if ! git status; then
         echo "llvm-project dir found but not a git repo, recloning"
-        cd $BUILDDIR
+        cd "$BUILDDIR"
         llvm_clone
     else
         echo "Existing llvm repo found, skipping clone"
         echo "Fetching new changes"
         llvm_pull
-        cd $BUILDDIR
+        cd "$BUILDDIR"
     fi
 else
     echo "cloning llvm project repo"
@@ -309,16 +308,16 @@ fi
 mkdir -p "$BUILDDIR/llvm-build"
 
 if [ -d "$BINUTILS_DIR"/ ]; then
-    cd $BINUTILS_DIR/
+    cd "$BINUTILS_DIR"/
     if ! git status; then
         echo "GNU binutils dir found but not a git repo, recloning"
-        cd $BUILDDIR
+        cd "$BUILDDIR"
         binutils_clone
     else
         echo "Existing binutils repo found, skipping clone"
         echo "Fetching new changes"
         binutils_pull
-        cd $BUILDDIR
+        cd "$BUILDDIR"
     fi
 else
     echo "cloning GNU binutils repo"
@@ -342,7 +341,7 @@ else
 fi
 cd "$OUT"
 
-LLVM_BIN_DIR=$(readlink -f $(which clang) | rev | cut -d'/' -f2- | rev)
+LLVM_BIN_DIR=$(readlink -f "$(which clang)" | rev | cut -d'/' -f2- | rev)
 
 OPT_FLAGS="-O3 -march=native -mtune=native -ffunction-sections -fdata-sections"
 OPT_FLAGS_LD="-Wl,-O3,--sort-common,--as-needed,-z,now -fuse-ld=$LLVM_BIN_DIR/ld.lld"
@@ -379,23 +378,23 @@ cmake -G Ninja -Wno-dev --log-level=NOTICE \
     -DLLVM_ENABLE_BACKTRACES=OFF \
     -DLLVM_ENABLE_WARNINGS=OFF \
     -DLLVM_ENABLE_LTO=Thin \
-    -DCMAKE_C_COMPILER=$LLVM_BIN_DIR/clang \
-    -DCMAKE_CXX_COMPILER=$LLVM_BIN_DIR/clang++ \
-    -DCMAKE_AR=$LLVM_BIN_DIR/llvm-ar \
-    -DCMAKE_NM=$LLVM_BIN_DIR/llvm-nm \
-    -DCMAKE_STRIP=$LLVM_BIN_DIR/llvm-strip \
-    -DLLVM_USE_LINKER=$LLVM_BIN_DIR/ld.lld \
-    -DCMAKE_LINKER=$LLVM_BIN_DIR/ld.lld \
-    -DCMAKE_OBJCOPY=$LLVM_BIN_DIR/llvm-objcopy \
-    -DCMAKE_OBJDUMP=$LLVM_BIN_DIR/llvm-objdump \
-    -DCMAKE_RANLIB=$LLVM_BIN_DIR/llvm-ranlib \
-    -DCMAKE_READELF=$LLVM_BIN_DIR/llvm-readelf \
-    -DCMAKE_ADDR2LINE=$LLVM_BIN_DIR/llvm-addr2line \
+    -DCMAKE_C_COMPILER="$LLVM_BIN_DIR"/clang \
+    -DCMAKE_CXX_COMPILER="$LLVM_BIN_DIR"/clang++ \
+    -DCMAKE_AR="$LLVM_BIN_DIR"/llvm-ar \
+    -DCMAKE_NM="$LLVM_BIN_DIR"/llvm-nm \
+    -DCMAKE_STRIP="$LLVM_BIN_DIR"/llvm-strip \
+    -DLLVM_USE_LINKER="$LLVM_BIN_DIR"/ld.lld \
+    -DCMAKE_LINKER="$LLVM_BIN_DIR"/ld.lld \
+    -DCMAKE_OBJCOPY="$LLVM_BIN_DIR"/llvm-objcopy \
+    -DCMAKE_OBJDUMP="$LLVM_BIN_DIR"/llvm-objdump \
+    -DCMAKE_RANLIB="$LLVM_BIN_DIR"/llvm-ranlib \
+    -DCMAKE_READELF="$LLVM_BIN_DIR"/llvm-readelf \
+    -DCMAKE_ADDR2LINE="$LLVM_BIN_DIR"/llvm-addr2line \
     -DLLVM_TOOL_CLANG_BUILD=ON \
     -DLLVM_TOOL_LLD_BUILD=ON \
     -DLLVM_CCACHE_BUILD=ON \
-    -DLLVM_PARALLEL_COMPILE_JOBS=$(nproc --all) \
-    -DLLVM_PARALLEL_LINK_JOBS=$(nproc --all) \
+    -DLLVM_PARALLEL_COMPILE_JOBS="$(nproc --all)" \
+    -DLLVM_PARALLEL_LINK_JOBS="$(nproc --all)" \
     -DCMAKE_C_FLAGS="$OPT_FLAGS" \
     -DCMAKE_ASM_FLAGS="$OPT_FLAGS" \
     -DCMAKE_CXX_FLAGS="$OPT_FLAGS" \
@@ -404,7 +403,7 @@ cmake -G Ninja -Wno-dev --log-level=NOTICE \
     -DCMAKE_SHARED_LINKER_FLAGS="$OPT_FLAGS_LD" \
     "$LLVM_PROJECT"
 
-ninja -j$(nproc --all) >/dev/null || (
+ninja -j"$(nproc --all)" >/dev/null || (
     echo "Could not build project!"
     exit 1
 )
@@ -458,26 +457,26 @@ cmake -G Ninja -Wno-dev --log-level=NOTICE \
     -DCOMPILER_RT_BUILD_XRAY=OFF \
     -DLLVM_ENABLE_TERMINFO=OFF \
     -DLLVM_ENABLE_LTO=Thin \
-    -DCMAKE_C_COMPILER=$STAGE1/clang \
-    -DCMAKE_CXX_COMPILER=$STAGE1/clang++ \
-    -DCMAKE_AR=$STAGE1/llvm-ar \
-    -DCMAKE_NM=$STAGE1/llvm-nm \
-    -DCMAKE_STRIP=$STAGE1/llvm-strip \
-    -DLLVM_USE_LINKER=$STAGE1/ld.lld \
-    -DCMAKE_LINKER=$STAGE1/ld.lld \
-    -DCMAKE_OBJCOPY=$STAGE1/llvm-objcopy \
-    -DCMAKE_OBJDUMP=$STAGE1/llvm-objdump \
-    -DCMAKE_RANLIB=$STAGE1/llvm-ranlib \
-    -DCMAKE_READELF=$STAGE1/llvm-readelf \
-    -DCMAKE_ADDR2LINE=$STAGE1/llvm-addr2line \
-    -DCLANG_TABLEGEN=$STAGE1/clang-tblgen \
-    -DLLVM_TABLEGEN=$STAGE1/llvm-tblgen \
+    -DCMAKE_C_COMPILER="$STAGE1"/clang \
+    -DCMAKE_CXX_COMPILER="$STAGE1"/clang++ \
+    -DCMAKE_AR="$STAGE1"/llvm-ar \
+    -DCMAKE_NM="$STAGE1"/llvm-nm \
+    -DCMAKE_STRIP="$STAGE1"/llvm-strip \
+    -DLLVM_USE_LINKER="$STAGE1"/ld.lld \
+    -DCMAKE_LINKER="$STAGE1"/ld.lld \
+    -DCMAKE_OBJCOPY="$STAGE1"/llvm-objcopy \
+    -DCMAKE_OBJDUMP="$STAGE1"/llvm-objdump \
+    -DCMAKE_RANLIB="$STAGE1"/llvm-ranlib \
+    -DCMAKE_READELF="$STAGE1"/llvm-readelf \
+    -DCMAKE_ADDR2LINE="$STAGE1"/llvm-addr2line \
+    -DCLANG_TABLEGEN="$STAGE1"/clang-tblgen \
+    -DLLVM_TABLEGEN="$STAGE1"/llvm-tblgen \
     -DLLVM_BUILD_INSTRUMENTED=IR \
     -DLLVM_BUILD_RUNTIME=OFF \
     -DLLVM_LINK_LLVM_DYLIB=ON \
     -DLLVM_VP_COUNTERS_PER_SITE=6 \
-    -DLLVM_PARALLEL_COMPILE_JOBS=$(nproc --all) \
-    -DLLVM_PARALLEL_LINK_JOBS=$(nproc --all) \
+    -DLLVM_PARALLEL_COMPILE_JOBS="$(nproc --all)" \
+    -DLLVM_PARALLEL_LINK_JOBS="$(nproc --all)" \
     -DCMAKE_C_FLAGS="$OPT_FLAGS" \
     -DCMAKE_ASM_FLAGS="$OPT_FLAGS" \
     -DCMAKE_CXX_FLAGS="$OPT_FLAGS" \
@@ -488,14 +487,14 @@ cmake -G Ninja -Wno-dev --log-level=NOTICE \
     "$LLVM_PROJECT"
 
 echo "Installing to $OUT/install"
-ninja install -j$(nproc --all) >/dev/null || (
+ninja install -j"$(nproc --all)" >/dev/null || (
     echo "Could not install project!"
     exit 1
 )
 
 STAGE2="$OUT/install/bin"
 PROFILES="$OUT/profiles"
-rm -rf "$PROFILES"/*
+rm -rf "${PROFILES:?}/"*
 echo "Stage 2: Build End"
 echo "Stage 2: PGO Train Start"
 
@@ -503,13 +502,13 @@ command -v aarch64-linux-gnu-as &>/dev/null || build_temp_binutils aarch64-linux
 command -v arm-linux-gnueabi-as &>/dev/null || build_temp_binutils arm-linux-gnueabi
 
 if [[ $USE_SYSTEM_BINUTILS_64 -eq 1 ]]; then
-    BINTUILS_64_BIN_DIR=$(readlink -f $(which aarch64-linux-gnu-as) | rev | cut -d'/' -f2- | rev)
+    BINTUILS_64_BIN_DIR=$(readlink -f "$(which aarch64-linux-gnu-as)" | rev | cut -d'/' -f2- | rev)
 else
     BINTUILS_64_BIN_DIR="$TEMP_BINTUILS_INSTALL/bin"
 fi
 
 if [[ $USE_SYSTEM_BINUTILS_32 -eq 1 ]]; then
-    BINTUILS_32_BIN_DIR=$(readlink -f $(which arm-linux-gnueabi-as) | rev | cut -d'/' -f2- | rev)
+    BINTUILS_32_BIN_DIR=$(readlink -f "$(which arm-linux-gnueabi-as)" | rev | cut -d'/' -f2- | rev)
 else
     BINTUILS_32_BIN_DIR="$TEMP_BINTUILS_INSTALL/bin"
 fi
@@ -537,7 +536,7 @@ make distclean defconfig \
     HOSTAR="$STAGE2"/llvm-ar \
     HOSTLD="$STAGE2"/ld.lld
 
-time make all -s -j$(nproc --all) \
+time make all -s -j"$(nproc --all)" \
     LLVM=1 \
     LLVM_IAS=1 \
     CC="$STAGE2"/clang \
@@ -574,7 +573,7 @@ make distclean defconfig \
     HOSTLD="$STAGE2"/ld.lld \
     CROSS_COMPILE=aarch64-linux-gnu-
 
-time make all -s -j$(nproc --all) \
+time make all -s -j"$(nproc --all)" \
     LLVM=1 \
     LLVM_IAS=1 \
     ARCH=arm64 \
@@ -595,7 +594,7 @@ time make all -s -j$(nproc --all) \
 
 # Merge training
 cd "$PROFILES"
-"$STAGE2"/llvm-profdata merge -output=clang.profdata *
+"$STAGE2"/llvm-profdata merge -output=clang.profdata ./*
 
 if [[ $BOLT_OPT -eq 0 ]]; then
     rm -rf "$TEMP_BINTUILS_INSTALL"
@@ -655,23 +654,23 @@ cmake -G Ninja -Wno-dev --log-level=NOTICE \
     -DCOMPILER_RT_BUILD_XRAY=OFF \
     -DLLVM_ENABLE_TERMINFO=OFF \
     -DLLVM_ENABLE_LTO=Thin \
-    -DCMAKE_C_COMPILER=$STAGE1/clang \
-    -DCMAKE_CXX_COMPILER=$STAGE1/clang++ \
-    -DCMAKE_AR=$STAGE1/llvm-ar \
-    -DCMAKE_NM=$STAGE1/llvm-nm \
-    -DCMAKE_STRIP=$STAGE1/llvm-strip \
-    -DLLVM_USE_LINKER=$STAGE1/ld.lld \
-    -DCMAKE_LINKER=$STAGE1/ld.lld \
-    -DCMAKE_OBJCOPY=$STAGE1/llvm-objcopy \
-    -DCMAKE_OBJDUMP=$STAGE1/llvm-objdump \
-    -DCMAKE_RANLIB=$STAGE1/llvm-ranlib \
-    -DCMAKE_READELF=$STAGE1/llvm-readelf \
-    -DCMAKE_ADDR2LINE=$STAGE1/llvm-addr2line \
-    -DCLANG_TABLEGEN=$STAGE1/clang-tblgen \
-    -DLLVM_TABLEGEN=$STAGE1/llvm-tblgen \
+    -DCMAKE_C_COMPILER="$STAGE1"/clang \
+    -DCMAKE_CXX_COMPILER="$STAGE1"/clang++ \
+    -DCMAKE_AR="$STAGE1"/llvm-ar \
+    -DCMAKE_NM="$STAGE1"/llvm-nm \
+    -DCMAKE_STRIP="$STAGE1"/llvm-strip \
+    -DLLVM_USE_LINKER="$STAGE1"/ld.lld \
+    -DCMAKE_LINKER="$STAGE1"/ld.lld \
+    -DCMAKE_OBJCOPY="$STAGE1"/llvm-objcopy \
+    -DCMAKE_OBJDUMP="$STAGE1"/llvm-objdump \
+    -DCMAKE_RANLIB="$STAGE1"/llvm-ranlib \
+    -DCMAKE_READELF="$STAGE1"/llvm-readelf \
+    -DCMAKE_ADDR2LINE="$STAGE1"/llvm-addr2line \
+    -DCLANG_TABLEGEN="$STAGE1"/clang-tblgen \
+    -DLLVM_TABLEGEN="$STAGE1"/llvm-tblgen \
     -DLLVM_PROFDATA_FILE="$PROFILES"/clang.profdata \
-    -DLLVM_PARALLEL_COMPILE_JOBS=$(nproc --all) \
-    -DLLVM_PARALLEL_LINK_JOBS=$(nproc --all) \
+    -DLLVM_PARALLEL_COMPILE_JOBS="$(nproc --all)" \
+    -DLLVM_PARALLEL_LINK_JOBS="$(nproc --all)" \
     -DCMAKE_C_FLAGS="$OPT_FLAGS" \
     -DCMAKE_ASM_FLAGS="$OPT_FLAGS" \
     -DCMAKE_CXX_FLAGS="$OPT_FLAGS" \
@@ -682,7 +681,7 @@ cmake -G Ninja -Wno-dev --log-level=NOTICE \
     "$LLVM_PROJECT"
 
 echo "Installing to $OUT/install"
-ninja install -j$(nproc --all) || (
+ninja install -j"$(nproc --all)" || (
     echo "Could not install project!"
     exit 1
 )
@@ -693,7 +692,7 @@ echo "Stage 3 Build: End"
 if [[ $BOLT_OPT -eq 1 ]]; then
     # Optimize final built clang with BOLT
     BOLT_PROFILES="$OUT/bolt-prof"
-    rm -rf $BOLT_PROFILES
+    rm -rf "$BOLT_PROFILES"
     mkdir -p "$BOLT_PROFILES"
     export PATH="$STAGE3:$BINTUILS_64_BIN_DIR:$BINTUILS_32_BIN_DIR:$STOCK_PATH"
     if [[ $CI -eq 1 ]]; then
@@ -703,8 +702,7 @@ if [[ $BOLT_OPT -eq 1 ]]; then
             exit 1
         )
     else
-        perf record -e cycles:u -j any,u -- sleep 1 &>/dev/null
-        if [[ $? == "0" ]]; then
+        if ! perf record -e cycles:u -j any,u -- sleep 1 &>/dev/null; then
             echo "Performing BOLT with sampling!"
             bolt_profile_gen "perf" || (
                 echo "Optimizing with BOLT failed!"
@@ -722,6 +720,6 @@ if [[ $BOLT_OPT -eq 1 ]]; then
 fi
 
 echo "Moving stage 3 install dir to build dir"
-mv $OUT/install $BUILDDIR/install/
+mv "$OUT"/install "$BUILDDIR"/install/
 echo "LLVM build finished. Final toolchain installed at:"
 echo "$BUILDDIR/install"
